Backend/pipeline/requirements.txt
pandas
yfinance
transformers
torch
python-dotenv